refactor: cleanup module structure a bit
Shuffle things around into a more rust-ish structure
