+++ rustfmt.toml
@@ -1,0 +1,12 @@
+# configuration for https://rust-lang.github.io/rustfmt/
+
+use_field_init_shorthand = true
+
+# unstable options. These require cargo +nightly fmt to use
+comment_width = 100 # generally more readable than 80
+format_code_in_doc_comments = true
+format_macro_matchers = true
+group_imports = "StdExternalCrate"
+imports_granularity = "Module" # generally leads to easier merges and shorter diffs
+normalize_doc_attributes = true # converts #[doc = "..."] to /// and #[doc(...)] to /// ...
+wrap_comments = true
