+++ sapphire-cli/src/cli/install.rs
