--- sapphire-cli/src/cmd/mod.rs
@@ -1,9 +1,0 @@
-// src/cmd/mod.rs
-// Declares the modules for each command implementation.
-
-pub mod info;
-pub mod install;
-pub mod search;
-pub mod uninstall;
-pub mod update;
-// Add other command modules here
